requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
dill
pyaml
mlflow
## -e .      ## this will go ahead and trigger the setup.py file called at the end of the project when the project is completed.